requirements.txt
# Core dependencies
anthropic>=0.25.0
requests>=2.31.0
python-dotenv>=1.0.0

# TTS Providers
elevenlabs>=0.2.0
cartesia>=1.1.0

# Document reading (optional but recommended)
python-docx>=1.1.0
PyPDF2>=3.0.0
python-pptx>=0.6.23
